==> analysis/03A_Ding-Mixture-HEK-3T3.Rmd <==
---
title: "Create Ding et al. datasets - Mixture 3T3/HEK"
output:
  workflowr::wflow_html:
    toc: false
editor_options:
  chunk_output_type: console
---

```{r}
suppressPackageStartupMessages({
  library(Seurat)
  library(ggplot2)
  library(patchwork)
  library(stringr)
  library(httr)
  library(XML)
})
set.seed(42)
theme_set(theme_classic())

FetchGEOFiles <- function(geo, download.dir = getwd(), download.files = FALSE, ...) {
  geo <- trimws(toupper(geo))
  geo_type <- substr(geo, 1, 3)
  url.prefix <- "https://ftp.ncbi.nlm.nih.gov/geo/"
  if (geo_type == "GSE") {
    url.prefix <- paste0(url.prefix, "series/")
  } else if (geo_type == "GSM") {
    url.prefix <- paste0(url.prefix, "samples/")
  } else if (geotype == "GPL") {
    url.prefix <- paste0(url.prefix, "platform/")
  }
  geo_prefix <- paste0(substr(x = geo, start = 1, stop = nchar(geo) - 3), "nnn")
  url <- paste0(url.prefix, geo_prefix, "/", geo, "/", "suppl", "/")
  response <- GET(url = url)
  html_parsed <- htmlParse(file = response)
  links <- xpathSApply(doc = html_parsed, path = "//a/@href")
  suppl_files <- as.character(grep(pattern = "^G", x = links, value = TRUE))
  if (length(suppl_files) == 0) {
    return(NULL)
  }
  file.url <- paste0(url, suppl_files)
  file_list <- data.frame(filename = suppl_files, url = file.url)
  if (download.files) {
    names(file.url) <- suppl_files
    download_file <- function(url, filename, ...) {
      message(paste0("Downloading ", filename, " to ", download.dir))
      download.file(url = url, destfile = file.path(download.dir, filename), mode = "wb", ...)
      message("Done!")
    }
    lapply(seq_along(file.url), function(y, n, i) {
      download_file(y[[i]], n[[i]], ...)
    },
    y = file.url, n = names(file.url)
    )
  }
  return(file_list)
}
```


```{r, message=FALSE, warning=FALSE}
download_dir <- here::here("data/raw_data/Ding")
dir.create(download_dir, showWarnings = F, recursive = T)
dir.create(here::here("data/rds_raw"), showWarnings = F, recursive = T)

geo_files <- FetchGEOFiles("GSE132044", download_dir, download.files = T)
geo_files
```


# Mixture

```{r}
mixture <- ReadMtx(mtx = "~/github/scRNA_NB_comparison/data/raw_data/Ding/GSE132044_mixture_hg19_mm10_count_matrix.mtx.gz",
  cells = "~/github/scRNA_NB_comparison/data/raw_data/Ding/GSE132044_mixture_hg19_mm10_cell.tsv.gz",
  features = "~/github/scRNA_NB_comparison/data/raw_data/Ding/GSE132044_mixture_hg19_mm10_gene.tsv.gz",
  feature.column = 1
)
```

```{r}
mixture[1:5, 1:5]
```

```{r}
replicate <- str_split_fixed(colnames(mixture), pattern = "\\.", n = 3)[, 1]
technology <- str_split_fixed(colnames(mixture), pattern = "\\.", n = 3)[, 2]
metadata <- data.frame(technology = technology, replicate = replicate)
rownames(metadata) <- colnames(mixture)

genes <- rownames(mixture)
species <- str_split_fixed(genes, pattern = "_", n = 4)[, 1]

hg19_genes <- species == "hg19"
mm10_genes <- species == "mm10"

mixture_hg19 <- mixture[hg19_genes, ]
mixture_mm10 <- mixture[mm10_genes, ]
```

```{r}
qplot(colSums(mixture_hg19), colSums(mixture_mm10))
```

```{r}
total_cell_umi <- data.frame(hg19_umi = colSums(mixture_hg19), mm10_umi = colSums(mixture_mm10))
total_cell_umi$hg19_over_mm10 <- total_cell_umi$hg19_umi / total_cell_umi$mm10_umi


human_cells <- rownames(total_cell_umi[total_cell_umi$hg19_over_mm10 >= 0.75, ])
mouse_cells <- rownames(total_cell_umi[total_cell_umi$hg19_over_mm10 < 0.25, ])

dim(total_cell_umi)
length(human_cells)
length(mouse_cells)
```

```{r}
mixture_hg19_humanonly <- mixture_hg19[, human_cells]
mixture_mm10_mouseonly <- mixture_mm10[, mouse_cells]

dim(mixture_hg19_humanonly)
dim(mixture_mm10_mouseonly)
```

```{r}
gene_ids_human <- rownames(mixture_hg19_humanonly)
gene_names_human <- make.unique(stringr::str_split_fixed(gene_ids_human, pattern = "_", n = 4)[, 4])
rownames(mixture_hg19_humanonly) <- gene_names_human

gene_ids_mouse <- rownames(mixture_mm10_mouseonly)
gene_names_mouse <- make.unique(stringr::str_split_fixed(gene_ids_mouse, pattern = "_", n = 4)[, 4])
rownames(mixture_mm10_mouseonly) <- gene_names_mouse
```

```{r}
mixture_hg19_humanonly[1:5, 1:5]
```

```{r}
mixture_mm10_mouseonly[1:5, 1:5]
```


```{r}
mixture_hg19 <- CreateSeuratObject(counts = mixture_hg19_humanonly, project = "Ding_scBenchmark_mixturehg19", min.cells=1, min.features = 1)
mixture_hg19 <- AddMetaData(object = mixture_hg19, metadata = metadata[human_cells, ])

mixture_mm10 <- CreateSeuratObject(counts = mixture_mm10_mouseonly, project = "Ding_scBenchmark_mixturemm10", min.cells=1, min.features = 1)
mixture_mm10 <- AddMetaData(object = mixture_mm10, metadata = metadata[mouse_cells, ])
```

```{r}
umi_techs <- c("10x-Chromium-v2", "CEL-Seq2", "Drop-seq", "inDrops", "sci-RNA-seq")
clean_named_techs <- c("ChromiumV2", "CEL-seq2", "Drop-seq", "inDrops", "sci-RNA-seq")

names(clean_named_techs) <- umi_techs
clean_named_techs
```

```{r}
`%notin%` <- Negate(`%in%`)

split_mixture_hg19 <- SplitObject(object = mixture_hg19, split.by = "technology")
names(split_mixture_hg19)


for (technology in names(split_mixture_hg19)) {
  if (technology %notin% umi_techs) next
  obj <- split_mixture_hg19[[technology]]
  obj_split <- SplitObject(object = obj, split.by = "replicate")

  clean_tech <- clean_named_techs[[technology]]


  for (sampletype in names(obj_split)) {
    seu_x <- obj_split[[sampletype]]
    seu_x[["percent.mt"]] <- PercentageFeatureSet(seu_x, pattern = "^MT-")

    saveRDS(
      seu_x,
      here::here("data/rds_raw", paste0("Ding-Human", sampletype, "__", clean_tech, ".rds"))
    )
  }
}
rm(split_mixture_hg19)
rm(mixture_hg19)
```

```{r}
split_mixture_mm10 <- SplitObject(object = mixture_mm10, split.by = "technology")
names(split_mixture_mm10)


for (technology in names(split_mixture_mm10)) {
  if (technology %notin% umi_techs) next
  obj <- split_mixture_mm10[[technology]]
  obj_split <- SplitObject(object = obj, split.by = "replicate")

  clean_tech <- clean_named_techs[[technology]]


  for (sampletype in names(obj_split)) {
    seu_x <- obj_split[[sampletype]]
    seu_x[["percent.mt"]] <- PercentageFeatureSet(seu_x, pattern = "^mt-")

    saveRDS(
      seu_x,
      here::here("data/rds_raw", paste0("Ding-Mouse", sampletype, "__", clean_tech, ".rds"))
    )
  }
}
```



```{r}
sessionInfo()
```
