==> analysis/09_Figure1.Rmd <==
---
title: "Figure 1"
output:
  workflowr::wflow_html:
    toc: false
editor_options:
  chunk_output_type: console
---

```{r setup, echo=FALSE}
suppressPackageStartupMessages({
  library(dplyr)
  library(ggplot2)
  library(ggpubr)
  library(ggridges)
  library(here)
  library(kableExtra)
  library(patchwork)
  library(RColorBrewer)
  library(readr)
  library(reshape2)
  library(scattermore)
  library(Seurat)
  library(sparseMatrixStats)
  library(xtable)
})
knitr::opts_chunk$set(warning = FALSE, message = FALSE)

`%notin%` <- Negate(`%in%`)
theme_set(theme_pubr(base_size = 10))
dir.create(here::here("output", "figures"), showWarnings = F)
clean_keys <- function(key) {
  gsub(
    pattern = "|\\)", replacement = "",
    x = gsub(pattern = " |\\(", replacement = "_", x = key)
  )
}

capitalize <- function(string) {
  substr(string, 1, 1) <- toupper(substr(string, 1, 1))
  string
}

global_labeller <- labeller(
  .default = capitalize, sample_name = label_wrap_gen(width = 15)
)

so_formatter <- function(x) {
  # x <- as.numeric(x)
  dplyr::case_when(
    x == "Full" ~ "All",
    as.numeric(x) == 377456 ~ "All",
    as.numeric(x) < 1e3 ~ as.character(x),
    as.numeric(x) < 1e6 ~ paste0(as.character(as.numeric(x) / 1e3), ",000"),
    as.numeric(x) < 1e9 ~ paste0(as.character(as.numeric(x) / 1e6), "M")
  )
}

PlotGLMResiduals <- function(fits_df, y = "ssr_pearson_normalized", ncol = 3, ylim = NULL) {
  colors <- brewer.pal(3, "Set2")
  names(colors) <- c("technical-control", "cell line", "heterogeneous")
  p <- ggplot(fits_df, aes_string("mean", y, color = "datatype")) +
    geom_scattermore(alpha = 0.5, pointsize = 1) +
    scale_x_continuous(trans = "log10", breaks = c(0.001, 0.1, 100), labels = MASS::rational) +
    facet_wrap(~sample_name, ncol = ncol, labeller = global_labeller) +
    scale_fill_manual(values = colors, name = "") +
    geom_hline(yintercept = 1, color = "black", linetype = "dashed") +
    ylab("Variance of quantile residuals") +
    xlab("Gene mean") +
    scale_color_manual(values = colors, name = "") +
    theme(
      legend.position = "bottom",
      legend.direction = "horizontal",
      legend.background = element_blank()
    ) +
    guides(col = guide_legend(ncol = ncol)) +
    guides(x = guide_axis(angle = 30)) +
    theme(legend.position = "none")

  if (!is.null(ylim)) {
    p <- p + ylim(0, 100)
  }
  return(p)
}

PlotGLMResiduals.sampled <- function(fits_df, y = "ssr_pearson_normalized", ncol = 3, ylim = NULL) {
  colors <- brewer.pal(3, "Set2")
  global_labeller1 <- labeller(
    .default = capitalize, sample_name = label_wrap_gen(width = 13)
  )

  names(colors) <- c("technical-control", "cell line", "heterogeneous")
  fits_df_subset <- fits_df[!is.na(fits_df[, y]), ]
  p <- ggplot(fits_df, aes_string("mean", y, color = "datatype")) +
    geom_scattermore(alpha = 0.5, pointsize = 1) +
    scale_fill_manual(values = colors, name = "") +
    scale_x_continuous(trans = "log10", breaks = c(0.001, 0.1, 100), labels = MASS::rational) +
    facet_wrap(~sample_name, ncol = ncol, labeller = global_labeller1) +
    geom_hline(yintercept = 1, color = "black", linetype = "dashed") +
    ylab("Variance of quantile residuals") +
    xlab("Gene mean") +
    scale_color_manual(values = colors, name = "") +
    theme(
      legend.position = "bottom",
      legend.direction = "horizontal",
      legend.background = element_blank()
    ) +
    guides(col = guide_legend(ncol = ncol)) +
    guides(x = guide_axis(angle = 30)) +
    theme(legend.position = "none")

  if (!is.null(ylim)) {
    p <- p + ylim(0, 100)
  }
  return(p)
}

PlotKDEPois <- function(fits_df) {
  p <- ggplot(fits_df, aes(mean, fill = gene_type)) +
    geom_histogram(aes(y = ..density..), position = "identity", alpha = 0.5) +
    geom_density(alpha = 0.6) +
    scale_x_continuous(trans = "log10", breaks = c(0.001, 0.1, 10, 1000), labels = MASS::rational) +
    scale_color_manual(values = brewer.pal(3, "Dark2"), name = "") +
    scale_fill_manual(values = brewer.pal(3, "Set1"), name = "Poisson goodness-of-fit test") +
    facet_wrap(~sample_name, scales = "free", labeller = global_labeller)
  return(p)
}

PlotNonPois <- function(frac_df) {
  colors <- brewer.pal(3, "Set2")
  names(colors) <- c("technical-control", "cell line", "heterogeneous")
  p <- ggplot(frac_df, aes(
    y = reorder(sample_name, nonpoisson_fraction, FUN = median),
    x = nonpoisson_fraction, color = datatype
  )) +
    geom_point(aes(size = median_cell_total_umi)) +
    scale_size(breaks = c(2000, 5000, 10000, 50000, 100000), labels = c("2K", "5K", "10K", "50K", "100K")) +
    scale_color_manual(values = colors, name = "Datatype") +
    scale_x_continuous(trans = "log10", breaks = c(0.001, 0.1, 0.25, 0.5, 0.75), labels = MASS::rational) +
    ylab("") +
    xlab("Non-poisson proportion") +
    theme(
      legend.position = "right", legend.direction = "vertical",
      legend.background = element_blank(),
    ) +
    labs(size = "Median UMI per cell") +
    guides(col = guide_legend(ncol = 1), size = guide_legend(ncol = 1)) +
    theme(legend.position = "right", legend.box = "vertical")
  return(p)
}


GetFits <- function(root_dir, qval_col, qval_threshold = 0.001) {
  fits <- sapply(datasets$key,
    FUN = function(x) {
      infile <- file.path(root_dir, x, "model_fit.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )

  cell_attrs_sampled <- sapply(datasets$key,
    FUN = function(x) {
      infile <- file.path(root_dir, clean_keys(x), "model_fit_cell_attr.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )

  cell_attrs_sampled_df <- bind_rows(cell_attrs_sampled, .id = "key")


  gene_attrs_sampled <- sapply(datasets$key,
    FUN = function(x) {
      infile <- file.path(root_dir, clean_keys(x), "model_fit_gene_attr.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )
  fits_df <- bind_rows(fits, .id = "key")
  gene_attrs_sampled_df <- bind_rows(gene_attrs_sampled, .id = "key")


  nonpoisson_fractions <- list()
  for (key in unique(fits_df$key)) {
    data_subset <- fits_df[fits_df$key == key, ]
    data_subset_nonpoisson <- data_subset[data_subset[, qval_col] < qval_threshold, ]
    cell_attrs_sampled_subset <- cell_attrs_sampled_df[cell_attrs_sampled_df$key == key, ]
    median_cell_total_umi <- median(cell_attrs_sampled_subset$total_umi)
    median_cell_avg_umi <- median(cell_attrs_sampled_subset$avg_umi)

    gene_attrs_sampled_subset <- gene_attrs_sampled_df[gene_attrs_sampled_df$key == key, ]
    median_gene_total_umi <- median(gene_attrs_sampled_subset$total_umi)
    median_gene_avg_umi <- median(gene_attrs_sampled_subset$avg_umi)

    nonpoisson_fractions[[key]] <- data.frame(
      nonpoisson_fraction = dim(data_subset_nonpoisson)[1] / dim(data_subset)[1],
      median_gene_total_umi = median_gene_total_umi,
      median_gene_avg_umi = median_gene_avg_umi,
      median_cell_total_umi = median_cell_total_umi,
      median_cell_avg_umi = median_cell_avg_umi
    )
  }
  frac_df <- bind_rows(nonpoisson_fractions, .id = "key")


  return(list(fits = fits_df, frac_df = frac_df, gene_attrs = gene_attrs_sampled_df, cell_attrs = cell_attrs_sampled_df))
}

ProcessFits <- function(fits_dfx, datasets, nb_column) {
  fits_df <- left_join(fits_dfx, datasets)
  fits_df <- fits_df %>% arrange(datatype)
  fits_df$sample_name <- factor(fits_df$sample_name, levels = datasets$sample_name)

  fits_df$gene_type <- "Pass"
  fits_df$gene_type[fits_df[, nb_column] < QVALUE_THRESHOLD] <- "Fail"
  return(fits_df)
}


GetMeanwiseNonPois <- function(fits_df) {
  labels <- c("<0.01", ">0.01", ">0.1", ">1", ">5", ">10", ">25", ">50", ">100", ">500")
  breaks <- c(min(fits_df$mean), 0.01, 0.1, 1, 5, 10, 25, 50, 100, 500, max(fits_df$mean))
  fits_df$mean_quantile <- cut(fits_df$mean,
    breaks = breaks,
    labels = labels
  )

  fits_df$mean_quantile <- factor(fits_df$mean_quantile, levels = labels)

  nonpoisson_fractions_meanwise <- list()

  keys <- c()
  for (key in unique(fits_df$key)) {
    keys <- c(key, keys)
    data_subset <- fits_df[fits_df$key == key, ]

    for (level in labels) {
      data_subset_level <- data_subset[data_subset$mean_quantile == level, ]
      data_subset_level_nonpois <- data_subset_level[data_subset_level$gene_type == "Fail", ]

      ssr_cols <- c("ssr_deviance_normalized", "ssr_pearson_normalized", "ssr_quantile_normalized")

      for (col in ssr_cols) {
        data_subset_level_nonpois[!is.finite(data_subset_level_nonpois[, col]), col] <- NA
      }

      ratio <- dim(data_subset_level_nonpois)[1] / dim(data_subset_level)[1]
      nonpoisson_fractions_meanwise[[paste(key, level)]] <- data.frame(
        key = key,
        mean_ssr_deviance_normalized = mean(data_subset_level_nonpois$ssr_deviance_normalized, na.rm = T),
        mean_ssr_pearson_normalized = mean(data_subset_level_nonpois$ssr_pearson_normalized, na.rm = T),
        mean_ssr_quantile_normalized = mean(data_subset_level_nonpois$ssr_quantile_normalized, na.rm = T),
        mean_quantile = level, nonpoisson_fraction = ratio, n_total_genes = dim(data_subset_level)[1]
      )
    }
  }
  nonpoisson_fractions_meanwise <- bind_rows(nonpoisson_fractions_meanwise)
  nonpoisson_fractions_meanwise$mean_quantile <- factor(nonpoisson_fractions_meanwise$mean_quantile, levels = labels)
  return(nonpoisson_fractions_meanwise)
}


PlotPoisVsUMI <- function(frac_df) {
  colors <- brewer.pal(3, "Set2")
  names(colors) <- c("technical-control", "cell line", "heterogeneous")
  p <- ggplot(frac_df, aes(median_cell_total_umi, nonpoisson_fraction)) +
    geom_point((aes(size = median_gene_avg_umi, fill = datatype)), colour = "black", alpha = 0.8, pch = 21) +
    facet_wrap(~datatype, labeller = global_labeller) +
    scale_size(breaks = c(0.03, 0.1, 0.5, 1)) +
    scale_fill_manual(values = colors, name = "Datatype") +
    xlab("Median total UMI per cell") +
    ylab("Non-poisson proportion") +
    guides(color = FALSE, fill = FALSE) +
    theme(
      legend.position = "bottom",
      legend.direction = "vertical",
      legend.background = element_blank(),
      legend.title = element_text(size = 9),
      legend.text = element_text(size = 9),
      #legend.box.background = element_rect(colour = "black")
    ) +
    guides(size = guide_legend(override.aes = list(color = "black", fill = "black"))) +
    scale_x_continuous(trans = "log10", breaks = c(1e3L, 1e4L, 1e5L), labels = c("1K", "10K", "100K")) +
    guides(col = guide_legend(ncol = 2), size = guide_legend(ncol = 3)) +
    labs(size = "Median UMI per gene")
  return(p)
}

PlotPoisVsUMI.cell_line <- function(frac_df) {
  colors <- brewer.pal(3, "Set2")
  names(colors) <- c("technical-control", "cell line", "heterogeneous")
  frac_df_subset <- frac_df[frac_df$datatype == "cell line", ]
  p <- ggplot(frac_df_subset, aes(median_cell_total_umi, nonpoisson_fraction, color=tissue)) +
    geom_point(size = 3, alpha = 0.8) +
    scale_color_manual(values = brewer.pal(3, "Set1"), name = "") +
    xlab("Median total UMI per cell") +
    ylab("Non-poisson proportion") +
    guides(color = TRUE, size = FALSE) +
    theme(
      legend.position = "bottom",
      legend.direction = "horizontal",
      legend.background = element_blank(),
      # legend.title = element_text(size = 12),
      # legend.text = element_text(size = 12),
      #legend.box.background = element_rect(colour = "black")
    ) +
    scale_x_continuous(trans = "log10", breaks = c(1e3L, 1e4L, 1e5L), labels = c("1K", "10K", "100K")) +
    guides(col = guide_legend(ncol = 2))
  return(p)
}

GetFits.sampled <- function(root_dir, qval_col, qval_threshold = QVALUE_THRESHOLD) {
  fits <- sapply(sampled_dataset_keys,
    FUN = function(x) {
      infile <- file.path(root_dir, x, "model_fit.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )

  cell_attrs_sampled <- sapply(sampled_dataset_keys,
    FUN = function(x) {
      infile <- file.path(root_dir, x, "model_fit_cell_attr.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )

  cell_attrs_sampled_df <- bind_rows(cell_attrs_sampled, .id = "key")


  gene_attrs_sampled <- sapply(sampled_dataset_keys,
    FUN = function(x) {
      infile <- file.path(root_dir, x, "model_fit_gene_attr.csv")
      if (file.exists(infile)) {
        return(read.csv(infile))
      } else {
        message(clean_keys(x))
      }
    },
    simplify = FALSE, USE.NAMES = TRUE
  )
  fits_df <- bind_rows(fits, .id = "key")
  gene_attrs_sampled_df <- bind_rows(gene_attrs_sampled, .id = "key")


  nonpoisson_fractions <- list()
  for (key in unique(fits_df$key)) {
    data_subset <- fits_df[fits_df$key == key, ]
    data_subset_nonpoisson <- data_subset[data_subset[, qval_col] < qval_threshold, ]
    cell_attrs_sampled_subset <- cell_attrs_sampled_df[cell_attrs_sampled_df$key == key, ]
    median_cell_total_umi <- median(cell_attrs_sampled_subset$total_umi)
    median_cell_avg_umi <- median(cell_attrs_sampled_subset$avg_umi)

    gene_attrs_sampled_subset <- gene_attrs_sampled_df[gene_attrs_sampled_df$key == key, ]
    median_gene_total_umi <- median(gene_attrs_sampled_subset$total_umi)
    median_gene_avg_umi <- median(gene_attrs_sampled_subset$avg_umi)

    nonpoisson_fractions[[key]] <- data.frame(
      nonpoisson_fraction = dim(data_subset_nonpoisson)[1] / dim(data_subset)[1],
      median_gene_total_umi = median_gene_total_umi,
      median_gene_avg_umi = median_gene_avg_umi,
      median_cell_total_umi = median_cell_total_umi,
      median_cell_avg_umi = median_cell_avg_umi
    )
  }
  frac_df <- bind_rows(nonpoisson_fractions, .id = "key")

  return(list(fits = fits_df, frac_df = frac_df, gene_attrs = gene_attrs_sampled_df, cell_attrs = cell_attrs_sampled_df))
}

ProcessFits.sampled <- function(fits_df, nb_column) {
  fits_df$gene_type <- "Poisson"
  fits_df$gene_type[fits_df[, nb_column] < QVALUE_THRESHOLD] <- "Non-poisson"
  return(fits_df)
}
```

## Read datassets

```{r}
datasets <- readr::read_csv(here::here("data", "datasets.csv"), col_types = readr::cols())
datasets$datatype <- factor(datasets$datatype, levels = c("technical-control", "cell line", "heterogeneous"))
datasets <- datasets %>% arrange(datatype)
```


## Read and process poisson GLM results

```{r}
QVALUE_THRESHOLD <- 1e-2
ncells <- "1000"
residual_type <- "quantile"
nb_column <- paste0("qval_ssr", "_", residual_type)
ssr_column <- paste0("ssr", "_", residual_type, "_normalized")
root_dir <- here::here("output/snakemake_output/poisson_glm_output", ncells)



fits <- GetFits(root_dir, qval_col = nb_column)
fits_dfx <- fits[["fits"]]
frac_dfx <- fits[["frac_df"]]
cell_attrs <- fits[["cell_attrs"]]
gene_attrs <- fits[["gene_attrs"]]

fits_df <- ProcessFits(fits_dfx, datasets, nb_column)
frac_df <- left_join(frac_dfx, datasets, by = "key")
```

## Fraction of non-poisson genes 

```{r}
frac_df_subset <- frac_df[, c(
  "sample_name", "datatype",
  "median_gene_avg_umi", "median_cell_total_umi", "nonpoisson_fraction"
)] %>% arrange(nonpoisson_fraction)
frac_df_subset <- frac_df_subset %>% mutate(across(where(is.numeric), round, 3))
kbl(frac_df_subset, booktabs = T) %>%
  kable_styling(latex_options = "striped")
dir.create(here::here("output", "tables"), showWarnings = F)
print(xtable(frac_df_subset, type = "latex",  digits=3), include.rownames = FALSE, file = here::here("output/tables/fraction_nonpoisson.tex"))
```

## Medium to high mean non-poisson genes

```{r}
medium_high_gene_pois <- list()
for (sample_name in unique(fits_df$sample_name)) {
  df <- fits_df[fits_df$sample_name == sample_name, ]

  med_abundance <- df[(df$mean > 1),]# & (df$mean <= 10), ]
  med_abundance_nonpois <- med_abundance[med_abundance$gene_type != "Pass", ]

  ratio <- dim(med_abundance_nonpois)[1] / dim(med_abundance)[1]

  medium_high_gene_pois[[sample_name]] <- data.frame(ratio = ratio, n_med_high_nonpoisson = dim(med_abundance_nonpois)[1], n_med_high = dim(med_abundance)[1], n_all = dim(df)[1])
}
medium_high_gene_pois <- bind_rows(medium_high_gene_pois, .id = "sample_name")

kbl(medium_high_gene_pois, booktabs = T) %>%
  kable_styling(latex_options = "striped")

print(xtable(medium_high_gene_pois,  digits=3), include.rownames = FALSE, digits=3, file = here::here("output/tables/medium_high_expr_nonpois_fraction.tex"))
```


## Plot fraction of non-poisson vs gene abundance

```{r fig1a,fig.width=8, fig.height=5}
nonpoismeanwise <- GetMeanwiseNonPois(fits_df)
nonpoismeanwise <- left_join(nonpoismeanwise, datasets, by = "key")
nonpoismeanwise_df <- nonpoismeanwise[complete.cases(nonpoismeanwise), ]
nonpoismeanwise_df$sample_name <- factor(nonpoismeanwise_df$sample_name, levels = datasets$sample_name)

nonpoismeanwise_filtered <- nonpoismeanwise_df[nonpoismeanwise_df$n_total_genes >= 10, ]


plot.dots <- ggplot(nonpoismeanwise_filtered, aes(sample_name, mean_quantile,
  color = nonpoisson_fraction,
  size = nonpoisson_fraction
)) +
  geom_point() +
  scale_color_viridis_c(
    direction = -1,
    breaks = c(0.25, 0.5, 0.75, 1),
    name = "Non-poisson fraction"
  ) +
  scale_size(
    name = "Non-poisson fraction",
    breaks = c(0.25, 0.5, 0.75, 1),
    range = c(0.5, 5)
  ) +
  theme_pubr(base_size = 10) +
  theme(legend.position = "right", legend.direction = "horizontal") +
  guides(color = guide_legend(), size = guide_legend()) +
  guides(x = guide_axis(angle = 90)) +
  xlab("") +
  ylab("Gene mean") +
  theme(legend.position = c(0.8, 1.06))

plot.dots
```


```{r}
nonpoismeanwise_filtered_subset <- nonpoismeanwise_filtered[, c("sample_name", "mean_quantile", "nonpoisson_fraction")]
nonpoismeanwise_filtered_subset2 <- tidyr::spread(nonpoismeanwise_filtered_subset, mean_quantile, nonpoisson_fraction)

kbl(nonpoismeanwise_filtered_subset2, booktabs = T) %>%
  kable_styling(latex_options = "striped")

print(xtable(nonpoismeanwise_filtered_subset2, type = "latex",  digits=3), digits=3, include.rownames = FALSE, file = here::here("output/tables/nonpoisson_meanwise.tex"))
```


## Plot residual variance


```{r fig1b}
sample_names <- c(
  "TechCtrl (inDrops)", "TechCtrl1 (ChromiumV1)",
  "HEK (ChromiumV3)",
  "PBMC (Smart-seq3)"
)

fits_df_subset <- fits_df[fits_df$sample_name %in% sample_names, ]
plot.glmresid4 <- PlotGLMResiduals(fits_df_subset, ssr_column, 4)
plot.glmresid4
```

## Distribution of means of poisson and non-poisson genes

```{r, fig.width=12, fig.height=10}
plot.kde <- PlotKDEPois(fits_df)
plot.kde
dir.create(here::here("output", "figures"), showWarnings = F, recursive = T)
ggsave(here::here("output", "figures", paste0("01_PoissonKDE_ncells_", ncells, "_residtype_", residual_type, ".pdf")), width = 12, height = 10)
```


## Plot fraction of non-poisson 

```{r}
plot.poisvsumi <- PlotPoisVsUMI.cell_line(frac_df)
plot.poisvsumix <- plot.poisvsumi + theme(legend.position = c(0.5, 1.0))
plot.poisvsumix
```

## Downsampling

```{r fig1d}
sample_name <- "PBMC__Smart-seq3"

root_dir <- here::here(paste0("output/snakemake_output/", sample_name, "_sampled_counts/poisson_glm_output/", ncells))


downsampled_umis <- c(
  10000, 7500, 5000,
  2000, 1000
)


names(downsampled_umis)<- c(
  "10,000", "7,500", "5,000",
  "2,000", "1,000"
)

sampled_dataset_keys <- paste0(sample_name, "_sampled_", downsampled_umis)

sampled_counts <- list()

for (name in downsampled_umis) {
  seu <- readRDS(here::here("data", "sampled_counts", paste0(sample_name, "_sampled_counts"), paste0(sample_name, "_sampled_", name, ".rds")))
  sampled_counts[[as.character(name)]] <- GetAssayData(seu, assay = "RNA", slot = "counts")
}

sampled <- readRDS(here::here("data", "rds_filtered", paste0(sample_name, ".rds")))
counts <- GetAssayData(sampled, assay = "RNA", slot = "counts")
counts_totalumis <- colSums2(counts)

generated_poissons <- list()

PredictPoisson <- function(counts_df) {
  predicted_poisson <- dpois(counts_df$counts, lambda = mean(counts_df$counts))
  return(predicted_poisson)
}

goi <- c("RPS19", "TPT1")
names(goi) <- goi

PredictPoisson <- function(counts_df) {
  predicted_poisson <- dpois(counts_df$counts, lambda = mean(counts_df$counts))
  return(predicted_poisson)
}

med_libsize <- "8,288" #(median(colSums2(counts)))

for (gene_name in names(goi)) {
  gene <- goi[[gene_name]]
  c1 <- "5000"
  c2 <- "1000"
  counts_goi_full <- counts[gene, ]
  counts_goi_full_subset <- counts_goi_full[counts_goi_full > 1]
  counts_goi_5k <- sampled_counts[[c1]][gene, ]
  counts_goi_1k <- sampled_counts[[c2]][gene, ]

  counts_df_full <- data.frame(counts = counts_goi_full_subset)
  counts_df_5k <- data.frame(counts = counts_goi_5k)
  counts_df_1k <- data.frame(counts = counts_goi_1k)

  labels <- c(
    paste0("Original  ", med_libsize, " UMI/cell"),
    paste0("Sampled ", so_formatter(as.integer(c1)), " UMI/cell"),
    paste0("Sampled ", so_formatter(as.integer(c2)), " UMI/cell")
  )
  counts_df_full$predicted_poisson <- PredictPoisson(counts_df_full)
  counts_df_full$sampletype <- labels[1]
  counts_df_5k$predicted_poisson <- PredictPoisson(counts_df_5k)
  counts_df_5k$sampletype <- labels[2]
  counts_df_1k$predicted_poisson <- PredictPoisson(counts_df_1k)
  counts_df_1k$sampletype <- labels[3]

  merged_counts <- rbind(counts_df_full, counts_df_5k, counts_df_1k)
  merged_counts$sampletype <- factor(merged_counts$sampletype, levels = labels)
  generated_poissons[[gene_name]] <- merged_counts
}
generated_poissons_df <- bind_rows(generated_poissons, .id = "gene")


plot.genespois <- ggplot(generated_poissons_df, aes(x = counts)) +
  geom_histogram(aes(y = ..density..), binwidth = 1, colour = "gray", fill = "white") +
  geom_line(aes(counts, predicted_poisson), color = "red") +
  facet_wrap(gene ~ sampletype, scales = "free", ncol = 3) +
  xlab("Counts") +
  ylab("Density") +
  scale_y_continuous(breaks = c(0.00, 0.10, 0.20))
plot.genespois
```

```{r, fig1e, fig.width=12, fig.height=4}
ncells <- "1000"
residual_type <- "quantile"
nb_column <- paste0("qval_ssr", "_", residual_type)
ssr_column <- paste0("ssr", "_", residual_type, "_normalized")





results_df <- GetFits.sampled(root_dir, nb_column)
fits_df.sampledx <- results_df$fits
frac_df.sampled <- results_df$frac_df

fits_df.sampled <- ProcessFits.sampled(fits_df.sampledx, nb_column)
fits_df.sampled$umi <- stringr::str_split_fixed(fits_df.sampled$key, pattern = "_", n = 5)[, 5]
frac_df.sampled$umi <- stringr::str_split_fixed(frac_df.sampled$key, pattern = "_", n = 5)[, 5]

fits_df.sampled$umi <- factor(fits_df.sampled$umi, levels = as.character(downsampled_umis), labels = paste0("Sampled ", names(downsampled_umis), " UMI/cell"))

frac_df.sampled$umi <- factor(frac_df.sampled$umi, levels = as.character(downsampled_umis), labels = paste0("Sampled ", names(downsampled_umis), " UMI/cell"))
fits_df.sampled$sample_name <- fits_df.sampled$umi
fits_df.sampled$datatype <- "heterogeneous"
fits_df.sampled$datatype <- factor(fits_df.sampled$datatype, levels = c("technical-control", "cell line", "heterogeneous"))
fits_df.sampled.subset <- fits_df.sampled[!is.na(fits_df.sampled[, ssr_column]), ]
```


```{r}
plot.glmresiddownsample <- PlotGLMResiduals.sampled(fits_df.sampled, y = ssr_column, ncol = 5) + ylab(paste0("Variance of ", residual_type, " residuals"))
plot.glmresiddownsample
```
```{r}
frac_df.sampled.subset <- frac_df.sampled[, c(
  "umi", "median_cell_total_umi", "median_gene_total_umi",
  "median_gene_avg_umi", "nonpoisson_fraction"
)]

kbl(frac_df.sampled.subset, booktabs = T) %>%
  kable_styling(latex_options = "striped")
print(xtable(frac_df.sampled.subset, type = "latex", digits=3), include.rownames = FALSE, file=here::here("output", "tables", "frac_nonpoisson_sampled.tex"))
```


```{r, fig.width=12, fig.height=12}
plot.poisvsumi <- PlotPoisVsUMI.cell_line(frac_df) + # theme(legend.position = "top")
theme(legend.position = c(0.5, 1.04))
layout <- "
AAAA
BBBC
DDDD
EEEE
"


plot.dots +
  plot.glmresid4 + plot.poisvsumi +
  plot.genespois + plot.glmresiddownsample +
  plot_layout(design = layout, tag_level = "new") + plot_annotation(tag_levels = "A") & theme(plot.tag = element_text(face = "bold"))
ggsave(here::here("output", "figures", "01_Figure1.pdf"), width = 12, height = 15)
```


```{r}
sessionInfo()
```
